refactor(abw): decompose itinerary composition
